Stylesheet and club changes
Added a javacsript script reference to the clubs.htm page. Formatted the
bootstrap.min stylesheet for viewing purposes.  Changed font color for
anchor tags to be white, get inherited when hovered over, change the
border color for portfolio items to a light grey, and changed the height
and width of a portflio item img. Got rid of hard coded file directory
for images and replaced it with a variable that gets the directory from
the database in Disc-Web-Development-Seminar.php.